fwk: Rework assertion logic
This commit modifies the assertion logic so that it is inlineable, and
uses more of the facilities provided by the compiler and standard
library for more robust error handling.
