+++ module/mock_psu/src/mod_mock_psu.c
@@ -304,7 +304,7 @@ static int mod_mock_psu_init(
     unsigned int element_count,
     const void *data)
 {
-    fwk_expect(data == NULL);
+    fwk_check(data == NULL);
 
     mod_mock_psu_ctx.elements =
         fwk_mm_calloc(element_count, sizeof(mod_mock_psu_ctx.elements[0]));
@@ -320,7 +320,7 @@ static int mod_mock_psu_element_init(
     struct mod_mock_psu_element_ctx *ctx;
     const struct mod_mock_psu_element_cfg *cfg = data;
 
-    fwk_expect(sub_element_count == 0);
+    fwk_check(sub_element_count == 0);
 
     ctx = mod_mock_psu_get_ctx(element_id);
